+++ CHANGELOG.md
@@ -8,10 +8,22 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 * Auto generation of row edit form.
 * Initial implementation of virtual scrolling.
 * Infinite scrolling.
+* Perfomance improvements.
 * Validation framework.
 * Persistence plugin.
 * Email, url cell editor.
 * Supporting of fetch in column filter.
+
+## [5.2.4] - 2018-02-16
+### Changed
+- Renamed `action-bar-core` to `action-bar`.
+- `layout="row"` refactored to `class="layout-row"`.
+- Export q-grid `common module` to make `q-grid-position` directive available for the end user.
+
+### Fixed
+- Url editor.
+- Text alingment in the file editor.
+- Navigation should trigger detect changes.
 
 ## [5.2.3] - 2018-02-12
 ### Fixed